third_party/WebKit/LayoutTests/scrollbars/scrollbar-layout-viewport-scrollwidth.html

<!DOCTYPE html>
<title>Verify scrollWidth for LayoutViewportScrollableArea</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<style>
    body, p{
        margin:0;
    }
    div{
        width:3000px;
    }
</style>
<body>
    <div>This page should be scrollable</div>
</body>
<script>
async_test(function(t) {
    window.onload = t.step_func_done(function() {
        assert_equals(document.scrollingElement.scrollWidth, 3000);
    });
});
</script>
